+++ Sources/RxLocationManager.swift
@@ -3,11 +3,16 @@
 //  RxLocationManager
 //
 //  Created by Anton Bronnikov on 03/09/2016.
-//  Copyright © 2016 northernForest. All rights reserved.
+//  Copyright © 2016 Anton Bronnikov. All rights reserved.
 //
 
+import CoreLocation
 import RxSwift
 
-class RxLocationManager {
+public final class RxLocationManager {
+
+    public static func standardLocationTracker() -> RxLocationTracker {
+        return RxLocationTracker()
+    }
 
 }