+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## [0.0.1](https://github.com/courteouselk/RxReachability/releases/tag/0.0.1)
 
-- Initial draft.
+- RxLocationTracker draft.
 - iOS and macOS targets.
 - MIT license (the project is in private repository, though, until it gets mature).